v1/requirements.txt
# IRIS v6.0 - Python Dependencies
# Förenklad och robust intelligensrapportering för svenska användare

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Client
aiohttp==3.10.5
httpx==0.27.2
requests==2.32.3

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0  # Async SQLite support
asyncpg==0.29.0  # Async PostgreSQL support
psycopg2-binary==2.9.9  # För PostgreSQL om önskad

# Cache
redis==5.2.1
hiredis==3.0.1  # Snabbare Redis

# AI och Language Processing
groq==0.11.0  # Groq Cloud SDK för Kimi K2
openai==1.52.0  # För xAI via OpenAI-kompatibel API
# xai-sdk==0.1.0  # Specifik xAI SDK när tillgänglig
transformers==4.46.0
sentence-transformers==3.1.1
spacy==3.8.1
langdetect==1.0.9

# Svenska NLP
# sv-core-news-sm  # Installeras med spacy download

# Validering och Konfiguration
python-dotenv==1.0.1
pyyaml==6.0.2
configparser==7.1.0

# Logging och Monitoring
structlog==24.4.0
loguru==0.7.2

# Security
cryptography==43.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Rate Limiting
slowapi==0.1.9
limits==3.13.0

# Async och Utilities
asyncio  # Built-in
aiofiles==24.1.0
async-timeout==4.0.3

# Data Processing
pandas==2.2.3
numpy==2.1.3
polars==1.12.0  # Snabbare än pandas för stora dataset

# Time and Date
pytz==2024.2
python-dateutil==2.9.0

# Testing (utveckling)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0  # Code coverage
pytest-timeout==2.3.1  # Test timeout
httpx==0.27.2  # För testing av FastAPI

# Environment och OS
psutil==6.1.0  # Systemövervakning

# JSON och Data
orjson==3.10.10  # Snabbare JSON
ujson==5.10.0

# Swedish-specific libraries
# Inga specifika svenska bibliotek behövs - använder standard NLP

# Development tools (valfritt)
black==24.10.0  # Code formatting
isort==5.13.2   # Import sorting
mypy==1.13.0    # Type checking

# Monitoring (produktion)
prometheus-client==0.21.0
sentry-sdk[fastapi]==2.17.0

# CORS och Middleware
python-multipart==0.0.12  # För form uploads

# Deployment
gunicorn==23.0.0  # WSGI server för produktion

# Health checks
health-check==1.3.3

# Version
# Frys alla versioner för reproducerbarhet
# pip freeze > requirements-frozen.txt